Cleaning up duplicates from ai summary merge
