+++ news-summary/requirements.txt
@@ -1,3 +1,3 @@
 # This is a list of packets for the issue
 # Every member needs to install using: pip install -r requirements.txt
--U google-genai
+google-genai
